Extra Refactoring
Code was too much for this class so tried refactoring it in another class and put it in a seperate library. Tried it only for loading and save methods to test if it works. Add the extra reference in the project. Next step is to extract all methods in the new library. Maybe too much for this project but it made sense.