Add Save functionality for University
Add Save functionality for University. Had to pass _university as parameter to the interface and the class that that implements the interface, because it is in a library. Bug: It saves only Univeristy and not students, grades, courses and scheduledcourses(this in not implemented yet).
Possible Fix: pass univeristy,students,grades,courses as parameters to save fuction in the library through the interface.
Note: Code seems very overwhelming at this point, not sure if this is a correct approach, interfaces should be simple