Add TransactionsController (has issues)
TransactionsController has issues first with precision since it only takes numbers with precision(4,2) and it doesn't make sens when we have to put the total price. Then it displays all pets and petfood and not for the pet selected.
Commented Out Date in TransactionCreateDto and TransactionEditDto since it is instantiated with datetime.now in Transaction constructor