Add MonthlyLedger with mock data
Note: Since the Transactions are not implemented correctly and because of lack of time for further coding, a hard coded table for monthly ledger with basic structure and a mock bar chart to make it more interesting for the user were added.
Bar chart was created in excel with random data for income and expenses. Total profit = income -expenses