Add PetReport mock page
Add PetReport mock page.
A mock table was added with hard coded data and a mock pie chart in an image at the bottom, created in excel with random values for total sold.
Note: Didn't have the time (and mental strength) to continue further coding and fix transactions and implement monthly ledger and pet report, so hard coded tables were used and some graphs created on excel to complete the Views. After further research, Asp.net supports the implementation of charts so they can be used to offer better user experience by visualizing table data. (However, in this exercise are not required and maybe should not be included)