Started Blazor tutorial from blazorschool.com
Started Blazor WebAssembly tutorial from blazorschool.com to try out Blazor. It's goal is to build an online store website.
The first step is to create a Product model and a ProductDisplay razor component. The Product.cs model was added to Session-26.Shared and the blazor component ProductDisplay.razor was added to to Sesssion-26.Client Shared folder. Using that was required was added to _Imports.razor