Add CartState and button buy ( Step 3a blazorschool tutorial)
Add CartState to Sates folder in client and button buy  and property DisplayBuyButton to ProductDisplay.razor. Then add property DisplayBuyButton when we use ProductDisplay in ProductList. In that way we create a cart service using state and then allow the user to add products to the cart when the buy button is clicked.
(Step 3a: BlazorSchool.com tutorial Wasm)