Add Session-27 Team Red - Car Service Center and FIX MonthlyLedger
Added Session-27, our second team project (team red) concerning  a Car Service Center in a Blazor Wasm solution.
Note: After a second, well rested look, monthly ledger was easy to be fixed. We have forgotten to add Parameters [Route("[controller]")] and [ApiController] and inheritance from ControllerBase and some usings. Also, inside foreach for manager and engineer there was no need to use  select  so this was removed too. Now monthlyledger works fine, but it is still needed to add data to it and fix prices calculations so that income can be calculated too.