Add seperate create page and dto, changes in repository
Add seperate Create Customer page and Dto for Customer, changes in repository in if statement inside Add() because Guid is already created in Customers constructor.
Note: Previously, when edit page was also create page, error 400: Bad Request was received in every attempt, because there was no Card Number in CustomerEditDto and the controller needed it. So, a CustomerCreateDto was made without the CardNumber and the controller was changed in Post to take CustomerCreateDto. Now, with this modification a new Customer is created successfully  and the card number is auto generated in Customer's constructor.