Modified Base Entity constructor and  CustomerRepository.
Modified BaseEntity in FuelSation.Model and removed Id generation from the constructor since it will be generated on add, when we put customer in the database. Also, modified CustomerRepository Add() method in FuelStation.EntityFramework, and changed the if statement, to check if entity.Id != default(Guid) and then throw exception.