Trying to fix Put in EmployeeController or EmployeeEdit page
Trying to fix Put in EmployeeController, getting error 400 constantly and it is not even getting in the controller. After debugging with breakpoints, the EmployeeEditDto Employee in EmployeeEdit  gets all updated values, but when calling PutAsJsonAsync it doesn't reach the controller. Connection refused and Error 400 is thrown. The updated Employee object seems correct.