Add UI Enhancements in Web Client
User Inteface Enhancements include:
- Created brand logo, slogan and app color palette
- Changed favicon
- Changed sidebar color
- Added logo to the sidebar
- Changed app title
- Changed color of the top row
- Changed nav item active color
- Added Epsilon Net Logo to the top right
NOTE: App name and slogan was one of the suggestions of ChatGPT.
Logo and slogan designs and color pallete was one of the suggestions of Logo.com
Background image in the main screen from Dribble.com
A big thanks to ChatGPT for the CSS styling at Index.razor
NOTE 2: UI Enhancements are not required but are a very important part of the user experience. There are some bugs, like the padding with the white space on the main image and some changes to the font sizes and button sizes are needed too, but these are not of priority at the time being.