Added CustomersForm with implementation for the datagrid - error in connection to the server
Added CustomersForm with implementation for the datagrid using method GetCustomers form CustomerService. Can't connect to server to check if it works.
NOTE: Keep getting error about the port. (Failed to bind to address {address number} address already in use).
Going to check it with netstat  -ano windows command to find the used ports in combination with the Process Identification Number (PID) in case I find the app in the task manager that uses it and then kill it.