Fixed previous error
Fixed previous error - "Unable to establish a connection because the destination computer actively refused it "by starting first the server from the command line. Noticed that the server was listening to the alternate port than the blazor client, also the port was  on http not https that i previously used. Also, changed the way httpclient was called and used ReadAsAsync method and not GetFromJsonAsync.
Now, GetCustomers() works fine.