Implemented ItemsForm (doesn't work completely)
Implemented ItemsForm (doesn't work completely), works for GetItems list, works for delete an item, doesn't work for add a new item and  update a new item because the forms CreateItem and UpdateItem don't work for combobox with item type. Also, not sure if the price and cost conversion works too.
NOTE 1: There are still many things to implement in the WinForms, but I am going to stop it here because the deadline has arrived.
NOTE 2: A big thank you to all our instructors for their support and guidance and to Epsilon Net for organizing this training program, giving entry level developers the chance to learn and evolve, learning essential job skills from experts.