Some fixes in UpdateTransactionForm
Some fixes in UpdateTransactionForm.  Used LoadCustomersAsync and LoadEmployeesAsync to get customers and employees lists. These are going to be used in customer and employee dropdowns.
Needs some extra fixes because in the dropdown the customer surname and customer name should be for the same id e.g when the user selects a customer surname, the name should appear in the dropdown for the same customer that has this id.