Implemented CreateTransactionForm
NOTE: There is still a lot of work to be done, now mostly on TransactionLine forms. Also, some business logic still is missing. In addition, a basic login should be implemented after the CRUD is complete and some validations.
For today, I am going to stop here, since I am already burnt out from this tough week and tired. It feels that I am close to complete it but I can't fully judge how much time it is going to take. I am going to leave it in this branch till I finish all the basic requirements when (and if) there is extra time. Thank you for this wonderful experience!